+++ NEWS.md
@@ -1,4 +1,4 @@
-# mmrm 0.2.2.9006
+# mmrm 0.2.2.9010
 
 ### New Features
 
@@ -12,8 +12,15 @@
 - Add empirical and empirical Jackknife adjusted coefficients covariance matrix.
   In addition, the argument `method` now only specifies the method used
   for the degrees of freedom, another argument `vcov` is added to specify the
-  method used to adjust the coefficients covariance matrix.
+  method used to adjust the coefficients covariance matrix. Empirical and empirical-Jackknife
+  covariance support residual and Satterthwaite degrees of freedom.
 - Add optional `tidymodels` framework support.
+
+### Bug Fixes
+
+- Previously `mmrm` fit follows the global option `na.action` and if it is set
+  other than "na.omit" error will happen. This is now fixed and `NA` values are
+  always removed prior to model fitting.
 
 # mmrm 0.2.2